Initial commit
Discerning repository readers may be wondering why this is the
initial commit, and where did all this code come from. Well, dear
reader, let me tell you a little bit about the history of this repo.

It began life as an idea for a piece of software very different from
what it would become. It has been a evening-and-weekend project for
a long time. And unfortunately, I tend to be pretty tired and lazy
on evening-and-weekends, and so I have -- at various points --
committed my personal API keys for certain services, re-worked vast
swathes of the code in single commits, and switched SCMs to
Mercurial, then Fossil, and now back to git. I also didn't imagine I
would ever make topiary public. I figured it would start and end its
life as a lonely side project.

Anyway, if you're from the far future and are curious about topiary
history before this commit, send me an email, and maybe I'll share
it if I'm feeling nostalgic.